fix: working copy
Still limited to the graph with 10 nodes, not 5K nor 100K

+++ source/FeederManager.h
@@ -23,7 +23,7 @@ struct intint {
 struct intintint {
     int father;
     int child;
-    int weight;
+    unsigned long int weight;
 };
 
 class feederManager {
